devils-codex/requirements.txt
# Core dependencies
fastapi==0.112.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.3
streamlit==1.32.2

# LLM and Vector DB
langchain==0.1.9
langchain-core==0.1.27
langgraph==0.0.41
langchain-openai==0.0.5
langchain-anthropic==0.0.10
langchain-community==0.0.19
chromadb==0.4.22
sentence-transformers==2.4.0
openai==1.14.0
anthropic==0.18.0

# Web tools and data processing
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.3
duckduckgo-search==4.1.1
pandas==2.2.0
numpy==1.26.4

# Testing and evaluation
pytest==7.4.4
ragas==0.1.3

# Deployment
gunicorn==21.2.0